# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPhOS 物理竞赛题生成 API",
    "description": "\nCPhOS 物理竞赛题全自动生成系统的后端 API。\n\n## 鉴权\n所有 `/api/**` 端点需在 `Authorization: Bearer <token>` 头中携带 token。\ntoken 由管理员通过 `/api/admin/tokens` 签发，明文仅创建时返回一次。\n\n## 异步工作流\n1. `POST /api/tasks` 提交任务，立即获得 `task_id`；\n2. 轮询 `GET /api/tasks/{task_id}` 直到 `status` 变为 `done` / `error` / `aborted`；\n3. 通过 `GET /api/tasks/{task_id}/artifacts` 与下载端点获取产物。\n",
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/llm/provider-kinds": {
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "列出已注册的服务商类型",
        "description": "返回服务商注册中心已注册的 kind 列表，供「新建服务商」表单的下拉框使用，\n新增 provider 后前端零改动。",
        "operationId": "list_provider_kinds_api_admin_llm_provider_kinds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderKindsResponse"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/llm/options": {
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "LLM / 应用设置聚合元数据",
        "description": "聚合返回服务商类型、Agent 角色与应用设置项元信息，使管理端表单完全由\n后端描述驱动（新增设置项 / 角色 / provider 无需改前端）。",
        "operationId": "get_llm_options_api_admin_llm_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMOptions"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/llm/providers": {
      "post": {
        "tags": [
          "llm-settings"
        ],
        "summary": "创建 LLM 服务商凭据",
        "description": "创建服务商凭据；``name`` 必须唯一。",
        "operationId": "create_provider_api_admin_llm_providers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "列出服务商",
        "description": "分页列出服务商（api_key 脱敏）。",
        "operationId": "list_providers_api_admin_llm_providers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProviderInfo_"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/llm/providers/{provider_id}": {
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "查看服务商",
        "operationId": "get_provider_api_admin_llm_providers__provider_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "llm-settings"
        ],
        "summary": "更新服务商",
        "description": "更新服务商；仅传入的字段会被修改（``api_key`` 留空表示保持不变）。",
        "operationId": "update_provider_api_admin_llm_providers__provider_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "llm-settings"
        ],
        "summary": "删除服务商",
        "description": "删除服务商；若仍被模型配置引用则拒绝。",
        "operationId": "delete_provider_api_admin_llm_providers__provider_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/llm/models": {
      "post": {
        "tags": [
          "llm-settings"
        ],
        "summary": "创建模型配置",
        "description": "创建模型配置；``name`` 唯一，``provider_id`` 必须存在。",
        "operationId": "create_model_config_api_admin_llm_models_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelConfigCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelConfigInfo"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "列出模型配置",
        "operationId": "list_model_configs_api_admin_llm_models_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ModelConfigInfo_"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/llm/agents": {
      "get": {
        "tags": [
          "llm-settings"
        ],
        "summary": "列出 Agent 绑定",
        "description": "列出全部权威 Agent 角色及其当前绑定的模型配置。",
        "operationId": "list_bindings_api_admin_llm_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentBindingInfo"
                  },
                  "type": "array",
                  "title": "Response List Bindings Api Admin Llm Agents Get"
                }
              }
            }
          },
          "401": {
            "description": "缺少或无效的 Bearer token。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "无权访问该资源（如非管理员 / 非任务归属者）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "与当前资源状态冲突（如任务非终态 / 编译进行中）。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定的机器可读错误码（如 not_found / forbidden / task_not_terminal）。",
            "examples": [
              "task_not_terminal"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "面向人类的错误说明。"
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "统一错误响应体。\n\n所有显式抛出的业务错误（``HTTPException`` / :class:`api.errors.ApiError`）均以\n此结构返回，便于前端按 ``code`` 做稳定的分支处理；``detail`` 为面向人类的提示。\n请求体校验错误（FastAPI 422）保留框架默认的结构化 ``{detail: [...]}`` 形态。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMOptions": {
        "properties": {
          "provider_kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Provider Kinds",
            "description": "已注册服务商类型清单。"
          },
          "agent_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agent Roles",
            "description": "权威 Agent 角色清单。"
          },
          "app_settings": {
            "items": {
              "$ref": "#/components/schemas/AppSettingSpec"
            },
            "type": "array",
            "title": "App Settings",
            "description": "运行期应用设置项及其类型 / 约束。"
          }
        },
        "type": "object",
        "required": [
          "provider_kinds",
          "agent_roles",
          "app_settings"
        ],
        "title": "LLMOptions",
        "description": "LLM / 应用设置的聚合元数据，驱动管理端表单按后端描述渲染。"
      },
...
      "ProviderKindsResponse": {
        "properties": {
          "kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kinds",
            "description": "服务商注册中心已注册的 kind（按字典序）。",
            "examples": [
              [
                "openai_compatible",
                "openrouter"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "kinds"
        ],
        "title": "ProviderKindsResponse",
        "description": "已注册的服务商类型（provider kind）清单。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "不透明 token，形如 `Bearer <token>`",
        "scheme": "bearer"
      }
    }
  }
}